+++ src/main/java/org/ojalgo/matrix/store/UpperHessenbergStore.java
@@ -66,7 +66,7 @@ final class UpperHessenbergStore<N extends Comparable<N>> extends ShadingStore<N
     }
 
     @Override
-    public Scalar<N> toScalar(final long row, final long col) {
+    public Scalar<N> toScalar(final int row, final int col) {
         if (row > col + 1) {
             return this.zero();
         }
