+++ src/main/java/org/ojalgo/matrix/store/UpperSymmetricStore.java
@@ -53,7 +53,7 @@ final class UpperSymmetricStore<N extends Comparable<N>> extends ShadingStore<N>
     }
 
     @Override
-    public Scalar<N> toScalar(final long row, final long col) {
+    public Scalar<N> toScalar(final int row, final int col) {
         if (row <= col) {
             return this.base().toScalar(row, col);
         }
